requirements.txt
pandas
dash
pyodbc
sqlalchemy
plotly
python-dotenv
matplotlib
numpy